+++ clikt/src/commonTest/kotlin/com/github/ajalt/clikt/core/ContextTest.kt
@@ -106,7 +106,7 @@ class ContextTest {
     @Test
     fun originalArgv() {
         class C : TestCommand()
-        class S: TestCommand() {
+        class S : TestCommand() {
             val opt by option()
             val args by argument().multiple()
             override fun run_() {
@@ -117,5 +117,17 @@ class ContextTest {
         }
 
         C().subcommands(S()).parse("s --opt o 1 2")
+    }
+
+    @Test
+    @JsName("assign_obj_through_context_builder")
+    fun `assign obj through context builder`() {
+        val foo = Foo()
+        val c = TestCommand()
+            .context { obj = foo }
+
+        c.parse("")
+
+        c.currentContext.obj shouldBeSameInstanceAs foo
     }
 }